# backend/cmd/pesto-server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A clipboard manager API with automatic monitoring",
        "title": "Pesto Clipboard API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/clipboard-items": {
            "get": {
                "description": "Get all clipboard items ordered by creation time",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clipboard"
                ],
                "summary": "Get all clipboard items",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ClipboardItem"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new clipboard item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clipboard"
                ],
                "summary": "Create a new clipboard item",
                "parameters": [
                    {
                        "description": "Clipboard Item",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ClipboardItem"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ClipboardItem"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/clipboard-items/content/{content_type}": {
            "get": {
                "description": "Get all clipboard items with a specific content type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clipboard"
                ],
                "summary": "Get clipboard items by content type",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Content Type",
                        "name": "content_type",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ClipboardItem"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/clipboard-items/current": {
            "get": {
                "description": "Get the current clipboard item (the most recently copied item)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clipboard"
                ],
                "summary": "Get the current clipboard item",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ClipboardItem"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/clipboard-items/{id}": {
            "get": {
                "description": "Get a specific clipboard item by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clipboard"
                ],
                "summary": "Get a clipboard item by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Clipboard Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ClipboardItem"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a clipboard item by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clipboard"
                ],
                "summary": "Delete a clipboard item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Clipboard Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ClipboardItem": {
            "description": "A clipboard item with its content and metadata",
            "type": "object",
            "properties": {
                "content": {
                    "description": "Content of the clipboard item",
                    "type": "string",
                    "example": "Hello, world!"
                },
                "content_type": {
                    "description": "Type of the content (e.g., text, image, url)",
                    "type": "string",
                    "example": "text"
                },
                "created_at": {
                    "description": "Creation timestamp",
                    "type": "string",
                    "example": "2025-01-01T00:00:00Z"
                },
                "deleted_at": {
                    "description": "Deletion timestamp (if deleted)",
                    "type": "string",
                    "example": "null"
                },
                "id": {
                    "description": "ID of the clipboard item",
                    "type": "integer",
                    "example": 1
                },
                "is_current": {
                    "description": "Whether this is the current clipboard item",
                    "type": "boolean",
                    "example": true
                },
                "updated_at": {
                    "description": "Last update timestamp",
                    "type": "string",
                    "example": "2025-01-01T00:00:00Z"
                }
            }
        }
    }
}